refactor: useGetVCard hook
