+++ components/lookup/person/ShareDialog.tsx
@@ -1,7 +1,7 @@
 // External libraries
 import va from "@vercel/analytics";
-import { useRouter } from "next/router";
 import { useTranslation } from "next-i18next";
+import { useRouter } from "next/router";
 
 // SK Components
 import {
@@ -18,8 +18,7 @@ import { DialogComponent } from "@/utils/types/common";
 import { Student, Teacher } from "@/utils/types/person";
 
 // Helpers
-import { getContactIsLinkable, getContactURL } from "@/utils/helpers/contact";
-import { getLocaleString } from "@/utils/helpers/i18n";
+import { useGetVCard } from "@/utils/helpers/contact";
 import { nameJoiner } from "@/utils/helpers/name";
 
 // Hooks
@@ -30,98 +29,17 @@ const ShareDialog: DialogComponent<{
 }> = ({ person, open, onClose }) => {
   const locale = useLocale();
   const { t } = useTranslation("lookup");
-  const { t: tx } = useTranslation("common");
 
   const router = useRouter();
+
+  const getVCard = useGetVCard();
 
   async function handleSaveVCard() {
     va.track("Share Person", {
       person: nameJoiner("en-US", person.name),
       method: "vCard",
     });
-
-    var vCard = new Blob(
-      [
-        [
-          // File header
-          `BEGIN:VCARD`,
-          `VERSION:3.0`,
-
-          // Name
-          `N:${nameJoiner(locale, person.name, undefined, {
-            firstName: false,
-          })};${nameJoiner(locale, person.name, undefined, {
-            lastName: false,
-          })};;${
-            person.role === "teacher"
-              ? t("people.dialog.share.saveVCard.segment.teacherPrefix")
-              : ""
-          };`,
-          `FN:${nameJoiner(locale, person.name, undefined, {
-            prefix: person.role === "teacher" ? "teacher" : false,
-          })}`,
-
-          // Birthday
-          `BDAY:${person.birthdate.split("-").join("")}`,
-
-          // Contacts
-          person.contacts
-            .map((contact, idx) => {
-              switch (contact.type) {
-                case "Phone":
-                  return `item${idx + 1}.TEL;type=CELL:${contact.value}`;
-                case "Email":
-                  return `item${idx + 1}.EMAIL;type=INTERNET:${contact.value}`;
-                default:
-                  if (getContactIsLinkable(contact))
-                    return [
-                      `item${idx + 1}.URL:${getContactURL(contact)}`,
-                      !["Website", "Other"].includes(contact.value) &&
-                        `item${idx + 1}.X-ABLabel:${
-                          {
-                            Facebook: tx("contact.facebook"),
-                            Line: tx("contact.line"),
-                            Instagram: tx("contact.instagram"),
-                            Discord: tx("contact.discord"),
-                          }[contact.type as string]
-                        }`,
-                    ]
-                      .filter((segment) => segment)
-                      .join("\n");
-                  break;
-              }
-            })
-            .filter((segment) => segment)
-            .join("\n"),
-
-          // Role within the school
-          person.role === "teacher" &&
-            [
-              `item1.ORG:${t(
-                "people.dialog.share.saveVCard.segment.org"
-              )};${getLocaleString(person.subjectGroup.name, locale)}`,
-              `item2.TITLE:${t("people.dialog.share.saveVCard.segment.title")}`,
-              person.classAdvisorAt &&
-                `NOTE:${t("people.dialog.share.saveVCard.segment.note", {
-                  number: person.classAdvisorAt.number,
-                })}`,
-            ]
-              .filter((segment) => segment)
-              .join("\n"),
-
-          // VCard metadata
-          `KIND:individual`,
-          `REV:${new Date().toISOString()}`,
-
-          // File footer
-          `END:VCARD`,
-        ]
-          .filter((segment) => segment)
-          .join("\n"),
-      ],
-      { type: "text/vcard;charset=utf-8" }
-    );
-
+    var vCard = getVCard(person);
     window.location.href = URL.createObjectURL(vCard);
     onClose();
   }
